renamed `entropy` to `rng` and moved the clock enabling of trng to main
