+++ chips/stm32u545/src/lib.rs
@@ -8,7 +8,7 @@ pub use stm32u5xx::{adc, chip, dma, exti, gpio, pwr, rcc, tim, usart};
 
 use cortexm33::{CortexM33, CortexMVariant};
 
-pub mod entropy;
+pub mod rng;
 
 #[cfg_attr(all(target_arch = "arm", target_os = "none"), used)]
 #[cfg_attr(all(target_arch = "arm", target_os = "none"), link_section = ".irqs")]
