+++ chips/stm32u5xx/src/lib.rs
@@ -8,12 +8,12 @@ pub mod adc;
 pub mod chip;
 pub mod dac;
 pub mod dma;
-pub mod entropy;
 pub mod exti;
 pub mod gpio;
 pub mod nvic;
 pub mod pwr;
 pub mod rcc;
+pub mod rng;
 pub mod tim;
 pub mod usart;
