+++ chips/stm32u5xx/src/rng.rs
